+++ Source/Tests/Fluxor.UnitTests/StoreTests/ThreadingTests/DispatchReentrancyTests/SupportFiles/CounterState.cs
@@ -1,0 +1,13 @@
+﻿namespace Fluxor.UnitTests.StoreTests.ThreadingTests.DispatchReentrancyTests.SupportFiles
+{
+	public class CounterState
+	{
+		public readonly int Counter;
+
+		public CounterState(int counter)
+		{
+			Counter = counter;
+		}
+	}
+}
+